+++ test/DataFilters.TestObjects/Person.cs
@@ -1,0 +1,16 @@
+using System;
+
+namespace DataFilters.TestObjects
+{
+    public class Person
+    {
+        public string Firstname { get; set; }
+
+        public string Lastname { get; set; }
+
+        public string Name { get; set; }
+
+        public DateTime BirthDate { get; set; }
+    }
+
+}