+++ test/DataFilters.UnitTests/SuperHero.cs
@@ -1,9 +1,0 @@
-﻿namespace DataFilters.UnitTests
-{
-    public class SuperHero
-    {
-        public string Name { get; set; }
-
-        public string Age { get; set; }
-    }
-}